server/requirements.txt
asn1crypto==0.24.0
astroid==1.6.3
backports.functools-lru-cache==1.5
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
cryptography==2.2.2
enum34==1.1.6
Faker==0.8.12
Flask==0.12.2
futures==3.2.0
idna==2.6
inflection==0.3.1
ipaddress==1.0.19
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pbr==4.0.2
phonenumbers==8.9.4
pycparser==2.18
PyJWT==1.6.1
pylint==1.8.4
pyOpenSSL==17.5.0
python-dateutil==2.7.2
python-dotenv==0.8.2
pytz==2018.4
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
stevedore==1.28.0
text-unidecode==1.2
twilio==6.11.0
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
Werkzeug==0.14.1
wrapt==1.10.11
